requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cloudinary==1.36.0
Django==4.2.7
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-jet-reboot==1.3.3
django-js-asset==2.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
PyJWT==2.8.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
urllib3==2.1.0
